pyproject.toml
[project]
name = "bedrock-logging-analytics"
version = "2.0.0"
description = "Bedrock invocation logging analytics - monitor token usage, costs and performance"
requires-python = ">=3.12"
dependencies = [
    "aws-cdk-lib>=2.100.0",
    "constructs>=10.0.0",
    "nicegui",
    "boto3",
]
